+++ src/otcurrentUIDialogBase.h
@@ -55,6 +55,7 @@ protected:
   virtual void OnCalendarShow(wxCommandEvent& event) { event.Skip(); }
   virtual void OnNow(wxCommandEvent& event) { event.Skip(); }
   virtual void OnSelectData(wxCommandEvent& event) { event.Skip(); }
+  virtual void OnSourceMode(wxCommandEvent& event) { event.Skip(); }
   virtual void About(wxCommandEvent& event) { event.Skip(); }
   virtual void OnPreferences(wxCommandEvent& event) { event.Skip(); }
 
@@ -62,10 +63,11 @@ public:
   wxTextCtrl* m_textCtrl1;
   wxChoice* m_choice1;
   wxTextCtrl* m_dirPicker1;
+  wxChoice* m_sourceModeChoice;
   wxButton* m_button8;
 
   otcurrentUIDialogBase(wxWindow* parent, wxWindowID id = wxID_ANY,
-                        const wxString& title = _("otcurrent_LTC_V.2.2"),
+                        const wxString& title = _("otcurrent_LTC_V.2.6_R"),
                         const wxPoint& pos = wxDefaultPosition,
                         const wxSize& size = wxSize(-1, -1),
                         long style = wxDEFAULT_DIALOG_STYLE | wxRESIZE_BORDER);